Prevent showing 0 for small numbers in dataframe (#10060)
## Describe your changes

If a number < 0.0001 is shown in a dataframe, it will be rounded to 0 in
our default formatting. This PR fixes it by showing at least the first
non-zero decimal of the number. E.g. if the number is `0.000000123` it
will format it to `0.0000001`.

## Testing Plan

- Added unit test for formatting function.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.