Add `height` parameter to `st.code` (#10080)
## Describe your changes

Adds a `height` parameter to `st.code`, which sets the height of the
code block and makes it scrollable if the content is larger than the
height.


https://github.com/user-attachments/assets/e19dcbc3-26d5-4b4f-a7d7-46491c5f685c


## GitHub Issue Link (if applicable)

Closes #7418

## Testing Plan

- Added one JS unit test to check for the correct height.
- Added two e2e tests, one with long code (enables scrolling) and one
with short code (makes the code block go beyond the length of the code).

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Johannes Rieke <johannes.rieke@snowflake.com>