Refactor arrow formatting utils to reduce reliance on pandas metadata (#9920)
## Describe your changes

Refactor the quiver formatting utils to be less dependent on pandas
metadata. This is part of the foundational dataframe improvements to
make our dataframe and table component able to handle raw arrow data (a
requirement for lazy loading).

## Testing Plan

- This change is not supposed to impact formatting behavior. There are
many existing e2e and unit tests covering the existing formatting
behavior.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.