Upgrade d3 dependencies to fix dependabot alert (#10109)
## Describe your changes

We have an open dependabot alert (high severity) related to d3-color
which cannot be auto-fixed:

https://github.com/streamlit/streamlit/security/dependabot/91

This PR upgrades all d3 dependencies in our frontend to resolve the
alert.

There are some small visual changes in the graphviz rendering (see
updated snapshots) and the graphviz component required a small fix.

## Testing Plan

- No logical changes -> no new tests required

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.