Allow markdown in `caption` of `st.image` (#10075)
## Describe your changes

This PR allows using markdown features for the `caption` parameter of
`st.image`. It has the same restrictions as widget labels, e.g. no
tables or horizontal rules.


## GitHub Issue Link (if applicable)

Closes #6808


## Testing Plan

- Added e2e tests in `label_markdown_test.py`. This is obviously not a
widget label, but I think it makes the most sense to test it there since
that file already contains all the strings for valid or invalid
Markdown. If that's not desired, I can also move the test to
`st_image_test.py`.
- Did not add unit tests because I think the e2e test should cover it
well enough and this is a pretty small feature. Plus, I don't think we
have unit tests for the other label markdown tests.
- Had to update a lot of snapshots because of subpixel movement. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Johannes Rieke <johannes.rieke@snowflake.com>