Fix server util unit test for local execution (#10131)
## Describe your changes

The `test_get_url` unit test can be a bit annoying when executing the
test suite locally. It fails if the port `8501` is used (e.g. by a dev
Streamlit instance). This changes the test to not use the automatic port
choosing behavior and instead enforces 8501 as the default port.

## Testing Plan

- Just a change in a unit test

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.