Changes needed for public notebooks to record protobufs (#10013)
## Describe your changes

These changes are needed for the public notebooks project and
accomplishes two main things:

1. Introduces `ForwardMsgList` allowing us to store all of the protobuf
messages as 1 protobuf message.
2. Adds a new lifecycle method that if set, will be called with the
message before it is enqueued.

But Nico! Why not utilize the newly added `ForwardMsgList` and avoid
this base64 tomfoolery? Great question! It boils down wanting to be able
to write the message as its enqueued. Why? For starters, it simplifies
the code a bit to be able to just append these base64 lines, but more
importantly, as we're recording these protobufs, we're watching this
file and waiting for it to stop getting updates and using that to
determine that the app is "done" rendering.


## Testing Plan
Added a unit test for this new method.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.