Use dynamic button label for button group (#9834)
## Describe your changes

Reuse the `DynamicButtonLabel` for pills & segmented control. Currently,
the usage of markdown in pills / segmented control is not limited ->
which isn't desired. Using `DynamicButtonLabel` applies the same limits
to the markdown as in all other button labels.

This also changes the starting emojis to be treated as icons instead of
text to have the same special icon spacing that we also have for emoji
icons on buttons.

## Testing Plan

- Only visual changes -> Update snapshots.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.