Activating linting rule for more consistent usage of built-in logging (#10133)
## Describe your changes

This PR activates the [flake8-logging-format
(G)](https://docs.astral.sh/ruff/rules/#flake8-logging-format-g) linting
rules from ruff to enforce better usage of the built-in logger. It also
moves all exceptions in logger output to `exc_info` which is the correct
way to handle exceptions via the built-in logger.

## Testing Plan

- No logical changes.
---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.