Print st.logo only on first page on the top (#10171)
## Describe your changes

Closes https://github.com/streamlit/streamlit/issues/10165

Fixes the printing layout of `st.logo`. Right now it shows on the side
(see https://github.com/streamlit/streamlit/issues/10165) but with this change it
will show on the top left and only on the first page.

This PR adds the Streamlit logo to the e2e version of the `hello_app`
test so that the logo printing is tested.

## GitHub Issue Link (if applicable)

## Testing Plan

- E2E Tests
- Add `st.logo` to the `hello_app` e2e test to ensure that printing is
working as expected

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.