Update the `typing-extensions` lower verion to >=0.4.4 to fix CI workflow & update snapshots (#10198)
## Describe your changes

Fixes our CI min-version test by updating the lower bound of
typing-extensions to 0.4.4. This is required by one of our transitive
dependencies. More info about the root cause in this PR:
https://github.com/python-jsonschema/referencing/pull/211

In addition to the issue with the dependency, there seems to be a second
independent issue that broke our CI pipeline:
The ubuntu version of our playwright e2e workflow got updated from 22 to
24 yesterday. This caused small changes in how the font is rendered,
which required most of our snapshots to be updated.


---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.